# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Banque",
        "description": "Documentation Swagger de l'API Banque",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes actifs",
                "description": "Récupère la liste des comptes non archivés avec filtrage avancé et pagination",
                "operationId": "getComptes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page (défaut: 1)",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page (défaut: 10, max: 100)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Epargne",
                                "Courant",
                                "Cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut du compte",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "inactif",
                                "bloque",
                                "suspendu"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche textuelle sur titulaire et numéro de compte",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "date_creation",
                                "solde",
                                "titulaire",
                                "numero_compte"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "Filtrer par ID client (pour les admins)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès"
                    },
                    "422": {
                        "description": "Paramètres de validation invalides"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte bancaire",
                "description": "Crée un nouveau compte bancaire pour un client. Si le client n'existe pas, il est créé automatiquement.",
                "operationId": "createCompte",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "soldeInitial",
                                    "devise",
                                    "client"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Type de compte",
                                        "type": "string",
                                        "enum": [
                                            "cheque",
                                            "epargne",
                                            "courant"
                                        ]
                                    },
                                    "soldeInitial": {
                                        "description": "Solde initial minimum 10 000",
                                        "type": "number",
                                        "minimum": 10000
                                    },
                                    "devise": {
                                        "description": "Devise du compte",
                                        "type": "string",
                                        "default": "XOF"
                                    },
                                    "client": {
                                        "required": [
                                            "titulaire",
                                            "email",
                                            "telephone",
                                            "adresse"
                                        ],
                                        "properties": {
                                            "titulaire": {
                                                "description": "Nom complet du titulaire",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Email du client",
                                                "type": "string",
                                                "format": "email"
                                            },
                                            "telephone": {
                                                "description": "Numéro de téléphone sénégalais",
                                                "type": "string"
                                            },
                                            "adresse": {
                                                "description": "Adresse du client",
                                                "type": "string"
                                            },
                                            "nci": {
                                                "description": "Numéro de carte d'identité",
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        },
...
        "/api/v1/comptes/{id}": {
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour les informations d'un compte bancaire",
                "description": "Met à jour le solde, le type ou le statut d'un compte bancaire",
                "operationId": "updateCompteBank",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "solde": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "statut": {
                                        "type": "string",
                                        "example": "actif"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "cheque"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour avec succès"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}